Refactor Game Logic
